# internal/adapter/http/docs/swagger.yaml
basePath: /api
definitions:
...
  http.splitTagMemberItemResponse:
    properties:
      id:
        type: string
      split_tag_id:
        type: string
      user_id:
        type: string
    type: object
  http.splitTagMemberResponse:
    properties:
      id:
        type: string
      split_tag_id:
        type: string
      user_id:
        type: string
      user_name:
        type: string
    type: object
  http.splitTagResponse:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      group_id:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: luis@example.com
    name: Luis Mackiewicz
  description: Backend for the Doheem expense management application
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/Luis-Mackiewicz/doheem-backend
  title: Doheem API
  version: "1.0"
paths:
...
  /api/groups/{groupId}/split-tags:
    get:
      consumes:
      - application/json
      description: List all split tags for a specific group
      parameters:
      - description: Group ID
        in: path
        name: groupId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/http.splitTagResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Group not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List split tags by group
      tags:
      - Split Tags
    post:
      consumes:
      - application/json
      description: Create a new split tag for a group
      parameters:
      - description: Group ID
        in: path
        name: groupId
        required: true
        type: string
      - description: Tag name
        in: body
        name: request
        required: true
        schema:
          properties:
            name:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/http.splitTagResponse'
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a split tag
      tags:
      - Split Tags
...
  /api/split-tags/{id}:
    delete:
      consumes:
      - application/json
      description: Permanently delete a split tag
      parameters:
      - description: Split Tag ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: group_id query param required
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Split tag not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a split tag
      tags:
      - Split Tags
  /api/split-tags/{id}/members:
    get:
      consumes:
      - application/json
      description: List all members assigned to a split tag
      parameters:
      - description: Split Tag ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/http.splitTagMemberResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Split tag not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List split tag members
      tags:
      - Split Tags
    post:
      consumes:
      - application/json
      description: Add a member to a split tag
      parameters:
      - description: Split Tag ID
        in: path
        name: id
        required: true
        type: string
      - description: User ID
        in: body
        name: request
        required: true
        schema:
          properties:
            user_id:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/http.splitTagMemberItemResponse'
        "400":
          description: Validation error
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Split tag not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Add split tag member
      tags:
      - Split Tags
  /api/split-tags/{id}/members/{userId}:
    delete:
      consumes:
      - application/json
      description: Remove a member from a split tag
      parameters:
      - description: Split Tag ID
        in: path
        name: id
        required: true
        type: string
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Remove split tag member
      tags:
      - Split Tags
...
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer <token>"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"